Moved files into pipeline folder
